Refactor import paths for consistency in MapView and related components
